+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 ### Removed
 * Resolving ambiguity using positional distance
+* `overload` is redundant with instance checking
 
 ## [1.12](https://pypi.org/project/multimethod/1.12/) - 2024-07-04
 ### Changed